API: use spec vars to avoid multiple initial edges
The strategy variables can be more than the
specification variables, for example the variable
`_goal` when using the interface
`tulip.interfaces.omega`. This change uses
specification variables instead of strategy variables.

Listing the set of variables using `list` avoids
a bug with Python 3, where `dict.keys()` returns
a view object [1], which is set-like [2].
In Python 2, `dict.keys()` returns a `list` [3].

[1] https://docs.python.org/3.8/library/stdtypes.html#dict.keys
[2] https://docs.python.org/3.8/library/stdtypes.html#dict-views
[3] https://docs.python.org/2/library/stdtypes.html#dict.keys