API: rm unused argument `use_cudd`
The API change in eee6daf15ab7c2c40d64a8642dc8336050a3cdb8
uses the module `dd.autoref` or the module `dd.cudd`
as selected in the module `omega.symbolic.fol`.

The selection in `omega.symbolic.fol` is when
importing modules. There is no direct mechanism
for changing this selection, so the argument
`use_cudd` would not be propagated.