update remaining calls to nodes_iter, edges_iter
as part of the transition to NetworkX 2.

Thanks to @johnyf (Ioannis Filippidis) for proposing the command

    ack --print0 -rl "edges_iter" | xargs -0 -L1 sed -i -e 's/edges_iter/edges/g'

and an entirely similar command with "nodes" instead of "edges".
Applying these commands yields this changeset, except for changes
under doc/, which require manual updating given changes to upstream
URLs and will instead appear in another documentation-only commit.