BUG: empty sys init when `qinit == '\A \A'`
The initial condition quantification `'\A \A'`
requires that `sys_init` be empty,
so the initial condition of the system generated
for transition graphs is conjoined to the
initial condition of the environment.

This ensures initialization of system variables
within their domains and so that the initial values
satisfy any other constraints.