CI: bump OS to `focal` (Ubuntu 20.04)
([Travis CI documentation][7]), because the binary
distribution of `gr1c` in the archive:
    https://github.com/tulip-control/gr1c/releases/download/v0.13.0/gr1c-0.13.0-Linux_x86-64.tar.gz
requires [GLIBC][10] 2.29, as confirmed (after unpacking the
distribution archive) with:

```shell
greadelf -s gr1c-0.13.0-Linux_x86-64/gr1c
```

where `greadelf` is the version of [`readelf`][8] that
is included in the [MacPorts package `binutils`][9].

The previous OS used in `.travis.yml` was
`bionic` ([Ubuntu 18.04][2]), which [has GLIBC 2.27][4].
In contrast, [`focal`][3] [has GLIBC 2.31][5].

Leaving the OS on CI unspecified would result in `xenial` being
installed ([as default][6]), which is [Ubuntu 16.04][1],
thus earlier than `bionic`.

Therefore, in order to use GLIBC 2.29, the CI environment
needs to change to use `focal`.

[1]: https://en.wikipedia.org/wiki/Ubuntu_version_history#Ubuntu_16.04_LTS_(Xenial_Xerus)
[2]: https://en.wikipedia.org/wiki/Ubuntu_version_history#Ubuntu_18.04_LTS_(Bionic_Beaver)
[3]: https://en.wikipedia.org/wiki/Ubuntu_version_history#Ubuntu_20.04_LTS_(Focal_Fossa)
[4]: https://packages.ubuntu.com/source/bionic/glibc
[5]: https://packages.ubuntu.com/source/focal/glibc
[6]: https://github.com/travis-ci/docs-travis-ci-com/blob/58c95e03d543058e817801ce7cabee92ce2507a3/user/reference/linux.md#overview
[7]: https://github.com/travis-ci/docs-travis-ci-com/blob/58c95e03d543058e817801ce7cabee92ce2507a3/user/reference/focal.md#using-ubuntu-2004-focal-fossa
[8]: https://en.wikipedia.org/wiki/Readelf
[9]: https://ports.macports.org/port/binutils/summary
[10]: https://en.wikipedia.org/wiki/GNU_C_Library