REL: rm from `requirements.txt` packages present in `install_requires`
The packages `networkx`, `omega`, `ply`, `scipy` are already
required via the parameter `install_requires` passed to the
function `setuptools.setup` in the script `setup.py`.

In general, there are two cases for what ranges of versions
for the requirements appear in `install_requires`:

1. the set of required versions contains only existing versions
  of packages. Assuming that the package is compatible with
  all these versions (as implied by these being required versions),
  it is unclear whether a `requirements.txt` would add
  any information to the information already available from
  the value of the parameter `install_requires`.

2. the set of required versions contains yet unreleased versions
  of packages, thus it is unknown whether the package is
  compatible with all of these versions (since this will be
  determined in the future). In this case, `requirements.txt`
  adds information to what is specified via the value of
  the parameter `install_requires`.

In this specific case, there are upper bounds on the versions of
`networkx` and `ply` in `install_requires` that are existing versions.
So for these packages case 1 from above applies.

The package `scipy` is not pinned in `requirements.txt`.
So for `scipy` the requirement in `requirements.txt` does not
add information to `install_requires`.

Only for the package `omega`, the set of versions specified in
`install_requires` can include future releases of `omega`.

Even so, the current upper bound (`omega < 0.4.0`) has been
written with the intention of keeping all such versions of
`omega` backwards-compatible, and releasing any
backwards-incompatible changes as version 0.4.0, or later.

Taking the above into consideration, it appears that these packages
do not need to be listed in `requirements.txt`.

In practice, if these packages are listed in `requirements.txt`,
then these constraints will need to be updated as the
requirements in `install_requires` change over time.
In other words, similar information is specified at two places.

It seems that the functionality used from packages whose required
versions are not upper-bounded is expected to remain unchanged
in the future. If there is an evident reason in the future to
pin in `requirements.txt` any package present in `install_requires`,
it can be added then to `requirements.txt`.