TST: `chdir` to `tests/` in `run_tests.py`, to write temp files in there
because tests run with `pytest` write generated files in the directory
where the script `run_tests.py` is invoked, even when `--rootdir` and
the configuration option `testpaths` are defined.

As a result of this change, the option `--where` is removed
(since the configuration file `pytest.init` is at a specific
place), and the code that changes `sys.path` is not needed now,
because `chdir` is called whenever `--outofsource` is given.

This change and the preceding commit are separate to keep
a snapshot of the script `run_tests.py` before this change.
A motivation is to have such a snapshot in the future,
in case `pytest` changes in a way that it will write files
generated by tests to the `--rootdir` directory. If so,
then reverting the present commit could be considered.