MNT: rm commented Python statement
This statement was introduced uncommented in
commit 6d32173478ec154b9ea96a9ab0670904843aa575,
which added the `ply`-based parser.

In that commit, the prime operator (`'`)
did not appear in any production rule of
the parser (only the precedence of the
prime operator was defined in the parser).
The lexer did recognize the prime operator.

The token for the prime operator was commented
in commit f68bb559aa444f31487d8201966c9e5ee67995f1.

The current token rule `PRIME`, which recognizes
the prime operator, was added in commit
4f3ec9d2556b2dc35b8bc9aeb52ab08a94dcd212.
The addition of a Python code statement in
that commit is not described by the commit's
message.

A production rule for the prime operator was
added to the parser in commit
b00937122e4c9fa54f828c1703e500849e63422e.

About the reason for why the token rule
was added in a commit whose title described
it differently, I created commit
b00937122e4c9fa54f828c1703e500849e63422e
about 15 minutes after creating commit
4f3ec9d2556b2dc35b8bc9aeb52ab08a94dcd212.

So the addition of the toke rule `t_PRIME`
was likely made in the same set of changes,
and the commit error happened while staging
the changes, i.e., in how the changes were
grouped into individual commits.

Moreover, commit 4f3ec9d2556b2dc35b8bc9aeb52ab08a94dcd212
was amended within 2 minutes of being created,
so perhaps the code statement was accidentally
staged when amending the commit.