API: uppercase Boolean literals in `lexyacc.Lexer.reserved`
This commit uppercases the literals `'false'` and `'true'`
in the keys of the dictionary stored in the
class attribute `tulip.spec.lexyacc.Lexer.reserved`.

The motivation for this change is TLA+ syntax,
and that internally in `tulip` the literals for
the Boolean constants were represented with
tokens that have as values uppercase strings.