API: rename positional parameters of `LabeledDiGraph` methods
to avoid naming coincidence of the positional parameters
and the keyword arguments of the methods:

- `tulip.transys.labeled_graphs.LabeledDiGraph.add_node`
- `tulip.transys.labeled_graphs.LabeledDiGraph.add_nodes_from`
- `tulip.transys.labeled_graphs.LabeledDiGraph.add_edge`

The new names of the parameters are the same with the
naming of the corresponding graph methods in `networkx == 2.6.3`.

The keyword parameters can include names of identifiers
that have been declared as variables by the user in
a logic context. The previous names of the positional
parameters of these methods were `n`, `nodes`, `u`, and `v`.
The new names are `node_for_adding`, `nodes_for_adding`,
`u_for_edge`, and `v_for_edge`, which are less likely
to occur in a specification.