REF: move class definitions to module scope, for AST node functors
The AST nodes are parametrized on two levels:

1. The `opmap` class-attribute defines a mapping of operators,
   which used almost only for non-nullary operator nodes
   (and also for Boolean constants; class `Bool`).

2. The instance attributes, which define what specific
   node this is, e.g., the node for the value `TRUE`,
   or the operator `+`.

The `opmap` parametrization is currently arranged at
the class level. So the functions `make_nodes()` and
`make_fol_nodes()` are class factories. The returned
value is a class that serves as a namespace that
contains classes, thus these functions resemble functors.

Flat is better than nested. So most of the
class definitions (i.e., the methods) is moved by
this commit to module scope. The module-scope classes
are now also inheriting from `typing.Protocol`.

The use of `typing.Protocol` aligns with the
factory-flavor of the module `spec.ast`.
This approach facilitates type hinting,
for example the return-type `NodesSpec` for
the function `tulip.spec.ast.make_nodes`.