REF: borrow imported module `matplotlib.pyplot` from `tulip.graphics`
Thus avoiding duplication of the `try` lines that
attempt to import `matplotlib`.

The call to the function `tulip.graphics._assert_pyplot()` is
in the function `tulip.abstract.plot.simulate2d()` made not
at the start of the function, but near the end, in order to
allow the `print` statement` of the function `simulate2d()` to
show its output.

The module attribute `tulip.graphics._plt` and the function
`tulip.graphics._assert_pyplot` are designated as internal.
However, these attributes are now used from outside the
module `tulip.graphics`. This usage to some extent contradicts
the leading underscore.

Nonetheless, the module `graphics` is part of the API of
the package `tulip`. Therefore, it is impossible to rename
`_plt` into `plt` and `_assert_pyplot` to `assert_pyplot`,
without at the same time adding these elements to `tulip`'s API.

In other words, the notion of "internal" is stratified:
- internal to a package (subpackages too)
- internal to a module

A way of implementing this stratification is to create
internal modules, i.e., modules whose name starts with
an underscore.

For example, by creating somewhere within `tulip`
a module called `_graphics` (name starts with an underscore),
then the API of the module `_graphics` is available to
other modules *within* `tulip` (assuming that the
new module's placement is as `tulip._graphics`),
but not to anywhere outside of `tulip`.

So the module `_graphics` would contain the attribute `plt`,
and a function named `assert_pyplot`, without these
elements becoming part of `tulip`'s API.

This kind of API internal to a package is designated
in commit titles with the prefix "_API",
in alignment with PEP 8 conventions.