BUG: cache upfront iterable, to avoid silently exhausting it
If while attempting to convert `big_iterable` to a `set`,
at the statement:

```python
big_iterable = set(big_iterable)
```

some item of `big_iterable` is unhashable,
then this will raise a `TypeError`.

The `except` section will then continue
the iteration, converting the rest of
the contents of `big_iterable` to
a `list`. This error is silent.

Moreover, converting `big_iterable` to a `set` is not needed,
whenever it is not already a `set`. It suffices that
`big_iterable` be a `Container`.