REF: spread `if` expression over multiple lines
to shorten line, and for readability.

This commit is described as "REF" because
it contains changes that are not blankspace.

So the abstract syntax tree definitely *does* change.

In Python blankspace defines what the code means,
so blankspace changes too, can change the AST.
Nonetheless, commits described as "blankspace changes"
are intended to leave unchanged the abstract syntax tree.

Blankspace commits that *do* change the
syntax tree are described as "REF".