API: rm unused parameter `accepting_states_type` from `__init__()`
Simple is better than complex:
use structural typing.
(And if the type of attribute `_accepting`
needs to be retrieved, use `type()`.)

Define attributes **only** if they need to
be used in an algorithm that you are implementing.
Thus, data structuring is avoided when not needed,
as recommended in the ["Epigrams on programming"](
    https://doi.org/10.1145/947955.1083808).