API: remove nested attribute `accepting`
Accepting states of automata are already accessible
using the attribute `accepting` of the automaton itself.
Flat is better than nested.

Simple is better than complex.
`pytype` raised an error when analyzing the function
`tulip.mvp._construct_weighted_product_automaton()`,
because the variable `fa` is an instance of the
class `tulip.transys.automata.WeightedFiniteStateAutomaton`,
but the attribute `fa.states.accepting` is defined
not in the class `tulip.transys.labeled_graphs.States`,
but later. This type-checking error supports the change.