REF: avoid `copy.deepcopy()`, just apply `list()`
This change is "REF" because of the type hint that
specifies either a `str` or an `Iterable[str]`.
So here a `list` as `Iterable` can only be a `list[str]`.

Strings are immutable, so `copy.deepcopy()` is
unnecessary in this case. For interned strings,
`copy.deepcopy()` is the identity mapping,
so it only adds to runtime. For strings that have not
been interned, `copy.deepcopy()` would unnecessarily
add to memory consumption.

Interestingly, in the type:

```python
Conjuncts = (
    str |
    _abc.Iterable[str])
```

a string too, is an iterable of strings.
So the choice of branch needs to be prioritized.