BUG: build parser tables in `setup.py`
Avoid `ModuleNotFoundError` in `setup.py`,
when attempting to build the parser tables.
Currently, the directory of `setup.py` is not
being added to `sys.path`, as it used to be
when running `python setup.py install`.

The result is that `import tulip` in `setup.py` raises
`ModuleNotFoundError`.
This error is circumvented by importing specific
modules from their files using `importlib`.

The next errors are due to the build environment
being different than the runtime environment:

- dependencies are not installed unless listed
  also in `[build-system]` requirements,
  which would cause duplication of `[project]`
  dependencies, and

- `import` from `tulip` does not work while `tulip`
  itself is being built, as it is not yet installed.

To avoid errors due to these `import` statements inside
the files loaded with `importlib`, `import` statements
are moved to inside functions.

The above are with:

- `pip == 25.3`
- `setuptools == 80.9.0`